Started implementing physical memory manager
* We calculate the header size, util size and real size of each memory
  map entry
* Added logging to debug port(I feel stupid for not knowing how to do it
  until now)
* Changed naming conventions from vnamespace_funcname to namespace_v_funcame for variadic functions
* Added a number of asm to c functions

+++ kernel/stdinc/math.h
@@ -22,12 +22,13 @@ int exp10i(int power);
  */
 int abs(int i);
 
-#define UNITS(n) (     (n)       )
-#define KILOS(n) (UNITS(n) / 1024)
-#define MEGS( n) (KILOS(n) / 1024)
-#define GIGS( n) (MEGS( n) / 1024)
-#define TERAS(n) (GIGS( n) / 1024)
-#define PETAS(n) (TERAS(n) / 1024)
-#define EXAS( n) (PETAS(n) / 1024)
+// cast cast cast cast cast cast cast
+#define UNITS(n) (uint64_t) (                ((uint64_t) n)                  )
+#define KILOS(n) (uint64_t) ((uint64_t) UNITS((uint64_t) n) / (uint64_t) 1024)
+#define MEGS( n) (uint64_t) ((uint64_t) KILOS((uint64_t) n) / (uint64_t) 1024)
+#define GIGS( n) (uint64_t) ((uint64_t) MEGS( (uint64_t) n) / (uint64_t) 1024)
+#define TERAS(n) (uint64_t) ((uint64_t) GIGS( (uint64_t) n) / (uint64_t) 1024)
+#define PETAS(n) (uint64_t) ((uint64_t) TERAS((uint64_t) n) / (uint64_t) 1024)
+#define EXAS( n) (uint64_t) ((uint64_t) PETAS((uint64_t) n) / (uint64_t) 1024)
 
 #endif
